feat(mindflow): v2.0.0 - MindFlow AI学习系统正式版
主要更新：
- MindFlow AI学习系统：对话/资源生成/云盘/学习画像
- 讯飞星火API集成（SSE流式响应）
- MinIO对象存储（学习资源文件管理）
- 云盘5类资源：DOC/MINDMAP/QUIZ/READING/CODE
- 学习画像：六维能力评估
- 资源批量导出ZIP功能
- CloudController/ResourceController/ProfileController完善

代码清理：
- 清理不必要文件（docs/, spark_demo/, AGENTS.md, release/, db_backup/）
- .gitignore完善
- UTF-16编码文件转换为UTF-8
- init.sql整合迁移脚本（file_key/file_size字段）

安全加固：
- CORS配置优化
- SSE超时配置
- 限流保护

版本：2.0.0 (versionCode 2)

+++ backend/src/main/java/com/chronoflow/backend/mindflow/entity/StudentProfile.java
@@ -50,9 +50,9 @@ public class StudentProfile {
     @TableField("error_types")
     private String errorTypes;
 
-    /** 画像版本号*/
+    /** 画像版本号 */
     @TableField("profile_version")
-    private Integer profileVersion;
+    private Short profileVersion;
 
     @TableField(value = "created_at", fill = FieldFill.INSERT)
     private LocalDateTime createdAt;
